+++ lizard_fewsjdbc/dtu.py
@@ -1,0 +1,11 @@
+"""Datetime utils"""
+
+from django.conf import settings
+from pytz import timezone
+
+def astimezone(timeseries, tz=timezone(settings.TIME_ZONE)):
+    """Localize an "aware" (i.e. non-naive) timeseries."""
+    for row in timeseries:
+        row['time'] = row['time'].astimezone(tz)
+    return timeseries
+